# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "knowledgebase",
    "description": "知识库文档摄入与检索 API",
    "version": "1.0.0"
  },
  "paths": {
    "/file-ingestion/upload": {
      "post": {
        "tags": [
          "file-ingestion"
        ],
        "summary": "Upload",
        "operationId": "upload_file_ingestion_upload_post",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_ingestion_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          },
          "500": {
            "description": "Ingestion failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/file-ingestion/delete": {
      "delete": {
        "tags": [
          "file-ingestion"
        ],
        "summary": "Delete File",
        "operationId": "delete_file_file_ingestion_delete_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileDeleteResponse"
                }
              }
            }
          },
          "500": {
            "description": "Delete failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/faq-ingestion/upsert": {
      "post": {
        "tags": [
          "faq-ingestion"
        ],
        "summary": "Upsert",
        "operationId": "upsert_faq_ingestion_upsert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FAQUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FAQIdResponse"
                }
              }
            }
          },
          "500": {
            "description": "FAQ upsert failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/faq-ingestion/update": {
      "put": {
        "tags": [
          "faq-ingestion"
        ],
        "summary": "Upsert Put",
        "operationId": "upsert_put_faq_ingestion_update_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FAQUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FAQIdResponse"
                }
              }
            }
          },
          "500": {
            "description": "FAQ update failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/faq-ingestion/{faq_id}": {
      "delete": {
        "tags": [
          "faq-ingestion"
        ],
        "summary": "Delete Faq",
        "operationId": "delete_faq_faq_ingestion__faq_id__delete",
        "parameters": [
          {
            "name": "faq_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Faq Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FAQIdResponse"
                }
              }
            }
          },
          "500": {
            "description": "FAQ delete failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/faq-ingestion/reindex": {
      "post": {
        "tags": [
          "faq-ingestion"
        ],
        "summary": "Reindex",
        "operationId": "reindex_faq_ingestion_reindex_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FAQReindexRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FAQReindexResponse"
                }
              }
            }
          },
          "500": {
            "description": "FAQ reindex failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/text-ingestion/upsert": {
      "post": {
        "tags": [
          "text-ingestion"
        ],
        "summary": "Upsert",
        "operationId": "upsert_text_ingestion_upsert_post",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextUpsertResponse"
                }
              }
            }
          },
          "500": {
            "description": "Text ingestion failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/text-ingestion/update": {
      "put": {
        "tags": [
          "text-ingestion"
        ],
        "summary": "Upsert Put",
        "operationId": "upsert_put_text_ingestion_update_put",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextUpsertResponse"
                }
              }
            }
          },
          "500": {
            "description": "Text update failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/text-ingestion/{doc_id}": {
      "delete": {
        "tags": [
          "text-ingestion"
        ],
        "summary": "Delete",
        "operationId": "delete_text_ingestion__doc_id__delete",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextDeleteResponse"
                }
              }
            }
          },
          "500": {
            "description": "Delete failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/healthz": {
      "get": {
        "summary": "Healthz",
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api-docs/openapi.json": {
      "get": {
        "tags": [
          "api-docs"
        ],
        "summary": "Api Openapi Json",
        "operationId": "api_openapi_json_api_docs_openapi_json_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api-docs/json-schema": {
      "get": {
        "tags": [
          "api-docs"
        ],
        "summary": "Api Json Schema",
        "operationId": "api_json_schema_api_docs_json_schema_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_file_ingestion_upload_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_file_ingestion_upload_post"
      },
      "FAQIdResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "FAQIdResponse"
      },
      "FAQReindexRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FAQUpsertRequest"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "FAQReindexRequest"
      },
      "FAQReindexResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "count"
        ],
        "title": "FAQReindexResponse"
      },
      "FAQUpsertRequest": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "markets": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Markets",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "question",
          "answer"
        ],
        "title": "FAQUpsertRequest"
      },
      "FileDeleteRequest": {
        "properties": {
          "doc_id": {
            "type": "string",
            "minLength": 1,
            "title": "Doc Id"
          }
        },
        "type": "object",
        "required": [
          "doc_id"
        ],
        "title": "FileDeleteRequest"
      },
      "FileDeleteResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "doc_id": {
            "type": "string",
            "title": "Doc Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "doc_id"
        ],
        "title": "FileDeleteResponse"
      },
      "FileUploadResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/FileUploadResultItem"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "count"
        ],
        "title": "FileUploadResponse"
      },
      "FileUploadResultItem": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "doc_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Id"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "blob_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blob Url"
          },
          "blob_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blob Key"
          },
          "toc_nodes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toc Nodes"
          },
          "toc_depth": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toc Depth"
          },
          "total_chunks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Chunks"
          },
          "chunking_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunking Mode"
          },
          "toc_outline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toc Outline"
          }
        },
        "type": "object",
        "required": [
          "filename"
        ],
        "title": "FileUploadResultItem"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ok"
          }
        },
        "type": "object",
        "title": "HealthResponse"
      },
      "TextDeleteResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "doc_id": {
            "type": "string",
            "title": "Doc Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "doc_id"
        ],
        "title": "TextDeleteResponse"
      },
      "TextUpsertRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "sourcefile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcefile",
            "default": "text.txt"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TextUpsertRequest"
      },
      "TextUpsertResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "doc_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Id"
          },
          "total_chunks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Chunks"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "TextUpsertResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}